# Switch ints that can't be negative to uints