--- a/model/internal/modeljson/http.go
+++ b/model/internal/modeljson/http.go
@@ -42,9 +42,9 @@ type HTTPRequestBody struct {
 type HTTPResponse struct {
 	Finished        *bool       `json:"finished,omitempty"`          // Non-ECS field.
 	HeadersSent     *bool       `json:"headers_sent,omitempty"`      // Non-ECS field.
-	TransferSize    *int64      `json:"transfer_size,omitempty"`     // Non-ECS field.
-	EncodedBodySize *int64      `json:"encoded_body_size,omitempty"` // Non-ECS field.
-	DecodedBodySize *int64      `json:"decoded_body_size,omitempty"` // Non-ECS field.
+	TransferSize    *uint64     `json:"transfer_size,omitempty"`     // Non-ECS field.
+	EncodedBodySize *uint64     `json:"encoded_body_size,omitempty"` // Non-ECS field.
+	DecodedBodySize *uint64     `json:"decoded_body_size,omitempty"` // Non-ECS field.
 	Headers         http.Header `json:"headers,omitempty"`           // Non-ECS field.
 	StatusCode      int         `json:"status_code,omitempty"`
 }