--- a/model/internal/modeljson/marshal_fastjson.go
+++ b/model/internal/modeljson/marshal_fastjson.go
@@ -575,7 +575,7 @@ func (v *Document) MarshalFastJSON(w *fastjson.Writer) error {
 	}
 	if v.DocCount != 0 {
 		w.RawString(",\"_doc_count\":")
-		w.Int64(v.DocCount)
+		w.Uint64(v.DocCount)
 	}
 	if v.Agent != nil {
 		w.RawString(",\"agent\":")
@@ -1080,7 +1080,7 @@ func (v *Event) MarshalFastJSON(w *fastjson.Writer) error {
 		} else {
 			w.RawString(prefix)
 		}
-		w.Int64(v.Severity)
+		w.Uint64(v.Severity)
 	}
 	if !v.SuccessCount.isZero() {
 		const prefix = ",\"success_count\":"
@@ -1503,7 +1503,7 @@ func (v *HTTPResponse) MarshalFastJSON(w *fastjson.Writer) error {
 		} else {
 			w.RawString(prefix)
 		}
-		w.Int64(*v.DecodedBodySize)
+		w.Uint64(*v.DecodedBodySize)
 	}
 	if v.EncodedBodySize != nil {
 		const prefix = ",\"encoded_body_size\":"
@@ -1513,7 +1513,7 @@ func (v *HTTPResponse) MarshalFastJSON(w *fastjson.Writer) error {
 		} else {
 			w.RawString(prefix)
 		}
-		w.Int64(*v.EncodedBodySize)
+		w.Uint64(*v.EncodedBodySize)
 	}
 	if v.Finished != nil {
 		const prefix = ",\"finished\":"
@@ -1584,7 +1584,7 @@ func (v *HTTPResponse) MarshalFastJSON(w *fastjson.Writer) error {
 		} else {
 			w.RawString(prefix)
 		}
-		w.Int64(*v.TransferSize)
+		w.Uint64(*v.TransferSize)
 	}
 	w.RawByte('}')
 	return nil
@@ -1836,7 +1836,7 @@ func (v *MessageAge) MarshalFastJSON(w *fastjson.Writer) error {
 	w.RawByte('{')
 	if v.Millis != nil {
 		w.RawString("\"ms\":")
-		w.Int64(*v.Millis)
+		w.Uint64(*v.Millis)
 	}
 	w.RawByte('}')
 	return nil
@@ -1910,7 +1910,7 @@ func (v *Histogram) MarshalFastJSON(w *fastjson.Writer) error {
 			if i != 0 {
 				w.RawByte(',')
 			}
-			w.Int64(v)
+			w.Uint64(v)
 		}
 		w.RawByte(']')
 	}
@@ -1936,7 +1936,7 @@ func (v *SummaryMetric) MarshalFastJSON(w *fastjson.Writer) error {
 	w.RawString("\"sum\":")
 	w.Float64(v.Sum)
 	w.RawString(",\"value_count\":")
-	w.Int64(v.Count)
+	w.Uint64(v.Count)
 	w.RawByte('}')
 	return nil
 }