--- a/model/modelpb/log_vtproto.pb.go
+++ b/model/modelpb/log_vtproto.pb.go
@@ -648,7 +648,7 @@ func (m *LogOriginFile) UnmarshalVT(dAtA []byte) error {
 				}
 				b := dAtA[iNdEx]
 				iNdEx++
-				m.Line |= int32(b&0x7F) << shift
+				m.Line |= uint32(b&0x7F) << shift
 				if b < 0x80 {
 					break
 				}