--- a/model/modelpb/process_vtproto.pb.go
+++ b/model/modelpb/process_vtproto.pb.go
@@ -594,7 +594,7 @@ func (m *ProcessThread) UnmarshalVT(dAtA []byte) error {
 				}
 				b := dAtA[iNdEx]
 				iNdEx++
-				m.Id |= int32(b&0x7F) << shift
+				m.Id |= uint32(b&0x7F) << shift
 				if b < 0x80 {
 					break
 				}