… codec/fullevent_test.go
… codec/utils_test.go
… input/elasticapm/docs/spec/rumv3/transaction.json
… input/elasticapm/docs/spec/v2/metricset.json
… input/elasticapm/docs/spec/v2/transaction.json
… input/elasticapm/internal/modeldecoder/v2/log_test.go
… input/elasticapm/internal/modeldecoder/v2/metricset_test.go
… input/otlp/logs_test.go
… input/otlp/metrics_test.go
… input/otlp/traces_test.go
… model/modelpb/apmevent.pb.json_test.go
… model/modelpb/event.pb.go
… model/modelpb/experience.pb.go
… model/modelpb/http.pb.go
… model/modelpb/http.pb.json_test.go
… model/modelpb/log.pb.go
… model/modelpb/message.pb.go
… model/modelpb/message.pb.json_test.go
… model/modelpb/message_vtproto.pb.go
… model/modelpb/metricset.pb.go
… model/modelpb/metricset.pb.json_test.go
… model/modelpb/metricset_vtproto.pb.go
… model/modelpb/process.pb.go
… model/modelpb/session.pb.go
… model/modelpb/session_vtproto.pb.go
… model/modelpb/transaction.pb.json_test.go
… model/modelpb/util_test.go
… model/proto/event.proto
… model/proto/experience.proto
… model/proto/http.proto