I am opening this PR as part of the #118 work.

```
Running benchstat
goos: darwin
goarch: arm64
pkg: github.com/elastic/apm-data/codec
                         │ bench-main  │            bench-branch            │
                         │   sec/op    │   sec/op     vs base               │
Encode/format=json-10      22.04µ ± 2%   22.19µ ± 3%       ~ (p=0.271 n=10)
Encode/format=vtproto-10   1.357µ ± 6%   1.359µ ± 2%       ~ (p=0.971 n=10)
Decode/format=json-10      16.23µ ± 7%   16.28µ ± 9%       ~ (p=0.684 n=10)
Decode/format=vtproto-10   6.423µ ± 3%   6.430µ ± 3%       ~ (p=0.912 n=10)
geomean                    7.472µ        7.496µ       +0.31%

                         │  bench-main  │             bench-branch              │
                         │   bytes/op   │   bytes/op    vs base                 │
Encode/format=json-10      6.169Ki ± 0%   6.169Ki ± 0%       ~ (p=1.000 n=10) ¹
Encode/format=vtproto-10   2.702Ki ± 0%   2.702Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                    4.083Ki        4.083Ki       +0.00%
¹ all samples are equal

pkg: github.com/elastic/apm-data/input/elasticapm/internal/decoder
                   │ bench-main  │         bench-branch          │
                   │   sec/op    │   sec/op     vs base          │
NDStreamDecoder-10   42.51n ± 7%   43.08n ± 7%  ~ (p=0.529 n=10)

pkg: github.com/elastic/apm-data/model/modelpb
                          │ bench-main  │         bench-branch          │
                          │   sec/op    │   sec/op     vs base          │
APMEventToJSON/to-json-10   11.62µ ± 3%   11.66µ ± 6%  ~ (p=0.684 n=10)
```