@dmathieu 2023-07-26
I have changed the code generation to keep the `elem < 0` check, as some values remain `nullable.Int`, due to them being from json, which doesn't know about signed integers.
And added a nolint to that bit of code generation to make staticcheck happy.